feat: move java directory validator and update imports
- Moved directory_validator.rs from java/validators to common/validators for shared use
- Created a new common/validators/mod.rs to re-export directory_validator
- Updated all Java command and service modules to import directory_validator from the new common location
- Removed redundant directory_validator module from java/validators
- Updated README and code comments to reflect the new shared validator structure
- Ensured all affected commands and services use the updated import paths for directory validation
