+++ src/commands/java/create_jpa_entity_id_field_command.rs
@@ -2,12 +2,14 @@ use std::path::Path;
 
 use crate::{
   commands::java::{
-    command_services::create_jpa_entity_id_field_service::run,
-    validators::directory_validator::validate_file_path_within_base,
+    services::create_jpa_entity_id_field_service::run,
     treesitter::types::id_field_config::IdFieldConfig,
     responses::file_response::FileResponse,
   },
-  common::response::Response,
+  common::{
+    response::Response,
+    validators::directory_validator::validate_file_path_within_base,
+  },
 };
 
 pub fn execute(
