+++ src/commands/java/create_jpa_repository_command.rs
@@ -2,14 +2,16 @@ use std::path::Path;
 
 use crate::{
   commands::java::{
-    command_services::create_jpa_repository_service::{run, run_with_manual_id},
-    validators::directory_validator::validate_file_path_within_base,
+    services::create_jpa_repository_service::{run, run_with_manual_id},
     responses::{
       create_jpa_repository_response::CreateJPARepositoryResponse,
       file_response::FileResponse,
     },
   },
-  common::response::Response,
+  common::{
+    response::Response,
+    validators::directory_validator::validate_file_path_within_base,
+  },
 };
 
 pub fn execute(
