+++ src/commands/java/get_java_files_command.rs
@@ -1,7 +1,7 @@
 use std::path::Path;
 
 use crate::{
-  commands::java::command_services::get_java_files_service::run,
+  commands::java::services::get_java_files_service::run,
   commands::java::treesitter::types::java_file_type::JavaFileType,
   commands::java::responses::{get_files_response::GetFilesResponse},
   common::response::Response,
