+++ src/commands/java/services/create_jpa_entity_basic_field_service.rs
