+++ src/commands/java/services/create_jpa_entity_enum_field_service.rs
