+++ src/commands/java/services/get_all_jpa_entities_service.rs
