+++ src/commands/java/services/get_all_packages_service.rs
