+++ src/commands/java/services/get_jpa_entity_info_service.rs
