+++ src/commands/java/validators/mod.rs
@@ -1,3 +1,2 @@
-pub mod directory_validator;
 pub mod java_class_name_validator;
 pub mod package_name_validator;
