+++ src/common/mod.rs
@@ -3,6 +3,7 @@ pub mod query;
 pub mod response;
 pub mod ts_file;
 pub mod utils;
+pub mod validators;
 
 #[cfg(feature = "ui")]
 pub mod ui;
